+++ sharding-scaling-bootstrap/src/main/java/info/avalon566/shardingscaling/Bootstrap.java
@@ -26,8 +26,8 @@ import info.avalon566.shardingscaling.core.config.ScalingConfiguration;
 import info.avalon566.shardingscaling.core.config.ScalingContext;
 import info.avalon566.shardingscaling.core.config.SyncConfiguration;
 import info.avalon566.shardingscaling.core.config.SyncType;
-import info.avalon566.shardingscaling.core.job.SyncTaskProgress;
-import info.avalon566.shardingscaling.core.job.ScalingJobController;
+import info.avalon566.shardingscaling.core.controller.SyncTaskProgress;
+import info.avalon566.shardingscaling.core.controller.ScalingJobController;
 import info.avalon566.shardingscaling.utils.RuntimeUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.log4j.PropertyConfigurator;
