+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/controller/ScalingJobController.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package info.avalon566.shardingscaling.core.job;
+package info.avalon566.shardingscaling.core.controller;
 
 import info.avalon566.shardingscaling.core.config.SyncConfiguration;
