+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/job/sync/executor/local/LocalSyncJobExecutor.java
@@ -18,8 +18,8 @@
 package info.avalon566.shardingscaling.core.job.sync.executor.local;
 
 import info.avalon566.shardingscaling.core.config.SyncConfiguration;
-import info.avalon566.shardingscaling.core.job.ReportCallback;
-import info.avalon566.shardingscaling.core.job.SyncTaskProgress;
+import info.avalon566.shardingscaling.core.controller.ReportCallback;
+import info.avalon566.shardingscaling.core.controller.SyncTaskProgress;
 import info.avalon566.shardingscaling.core.job.sync.SyncTask;
 import info.avalon566.shardingscaling.core.job.sync.executor.Event;
 import info.avalon566.shardingscaling.core.job.sync.executor.Reporter;
